requirements.txt
pycryptodome==3.20.0
pywin32==306
customtkinter==5.2.2
